+++ src/parameters/default.jl
@@ -21,7 +21,7 @@ const DEFAULT_PATHOGEN_TYPE = PathogenType(
 
 const DEFAULT_RESPONSE_TYPE = ResponseType(
     "Default",
-    0.0,
+    # 0.0,
     # (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     # (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
     # (hos_g::String, imp_g::String, mat_g::String, pat_g::String) -> 1.0,
@@ -77,12 +77,12 @@ const DEFAULT_POPULATION_TYPE = PopulationType(
     # 1.0,
     # 1.0,
     # 0.0,
-    0.0,
+    # 0.0,
     true,
-    0.0,
+    # 0.0,
     (h_1::String, h_2::String) -> true,
-    g::String -> 1.0, # takes seq argument, returns Float64
-    g::String -> 1.0, # takes seq argument, returns Float64
+    # g::String -> 1.0, # takes seq argument, returns Float64
+    # g::String -> 1.0, # takes seq argument, returns Float64
     SA[ # order defined in COEFFICIENTS
         0.0, 0.0, 0.0,
         0.0, 0.0, 0.0,